# docs/dashboard-openapi.yaml
# AUTO-GENERATED — DO NOT EDIT BY HAND.
# Source of truth: apps/tenant-dashboard/src/app/api/**/route.ts (withApi + Zod).
# Regenerate via `yarn workspace tenant-dashboard gen:openapi`. CI enforces no drift.
components:
  parameters: {}
  responses:
    InternalError:
      content:
        application/json:
          schema:
            properties:
              error:
                additionalProperties: true
                properties:
                  code:
                    type: string
                  details:
                    additionalProperties: true
                    type: object
                  message:
                    type: string
                required:
                  - code
                  - message
                type: object
            required:
              - error
            type: object
      description: Unexpected server error. The client has not malformed the request; the failure is on our side and has been logged. Safe to retry with backoff.
  schemas: {}
  securitySchemes:
    SessionCookie:
      description: Supabase session cookie set on sign-in. Dashboard-UI-only — machine clients should use the Gateway API (api.agentmark.co) with an API key instead.
      in: cookie
      name: sb-access-token
      type: apiKey
info:
  description: Internal HTTP API that the OuterLayer dashboard web app uses for tenant-scoped reads/writes. This surface is served by the same Next.js app as the UI and authenticates via Supabase session cookie. Not a public API — for programmatic access use the Gateway API at api.agentmark.co.
  title: OuterLayer Dashboard API
  version: "1.0"
openapi: 3.0.3
paths:
  /api/orgs/{orgName}/apps/{appId}/agents/blob/{sha256}:
    get:
      operationId: get-agents-blob
      parameters:
        - in: path
          name: orgName
          required: true
          schema:
            minLength: 1
            type: string
        - in: path
          name: appId
          required: true
          schema:
            minLength: 1
            type: string
        - in: path
          name: sha256
          required: true
          schema:
            pattern: ^[a-f0-9]{64}$
            type: string
        - in: query
          name: appId
          required: true
          schema:
            type: string
        - in: query
          name: token
          required: true
          schema:
            minLength: 1
            type: string
      responses:
        "200":
          description: Image bytes.
        "403":
          content:
            application/json:
              schema:
                properties:
                  error:
                    additionalProperties:
                      nullable: true
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    type: object
                required:
                  - error
                type: object
          description: Missing, altered, or expired image token.
        "404":
          content:
            application/json:
              schema:
                properties:
                  error:
                    additionalProperties:
                      nullable: true
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    type: object
                required:
                  - error
                type: object
          description: Not found.
        default:
          $ref: "#/components/responses/InternalError"
      summary: Agent image blob (content-addressed)
      tags:
        - Agents
  /api/orgs/{orgName}/apps/{appId}/onboarding/checklist:
    get:
      description: Returns the onboarding setup state (checklist booleans + git provider) — drives the getting-started checklist, the setup card gate, and the soft banner.
      operationId: onboarding-checklist
      parameters:
        - in: path
          name: orgName
          required: true
          schema:
            minLength: 1
            type: string
        - in: path
          name: appId
          required: true
          schema:
            minLength: 1
            type: string
        - in: query
          name: appId
          required: true
          schema:
            minLength: 1
            type: string
      responses:
        "200":
          description: "`{ hasApiKey, hasTrace, hasTeammate, hasGitConnection, hasRepoLinked, provider }`."
        "400":
          content:
            application/json:
              schema:
                properties:
                  error:
                    additionalProperties:
                      nullable: true
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    type: object
                required:
                  - error
                type: object
          description: Invalid query params.
        "401":
          content:
            application/json:
              schema:
                properties:
                  error:
                    additionalProperties:
                      nullable: true
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    type: object
                required:
                  - error
                type: object
          description: Not authenticated.
        default:
          $ref: "#/components/responses/InternalError"
      summary: Getting-started checklist status
      tags:
        - Onboarding
  /api/orgs/{orgName}/has-traces:
    get:
      description: "Returns `{ hasTraces: boolean }` — used to drive the onboarding placeholder in the dashboard."
      operationId: has-traces
      parameters:
        - in: path
          name: orgName
          required: true
          schema:
            minLength: 1
            type: string
        - in: query
          name: appId
          required: true
          schema:
            minLength: 1
            type: string
      responses:
        "200":
          description: "`{ hasTraces: boolean }`."
        "400":
          content:
            application/json:
              schema:
                properties:
                  error:
                    additionalProperties:
                      nullable: true
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    type: object
                required:
                  - error
                type: object
          description: Invalid query params.
        "401":
          content:
            application/json:
              schema:
                properties:
                  error:
                    additionalProperties:
                      nullable: true
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    type: object
                required:
                  - error
                type: object
          description: Not authenticated.
        default:
          $ref: "#/components/responses/InternalError"
      summary: Onboarding existence check
      tags:
        - Analytics
security:
  - SessionCookie: []
servers:
  - description: Production dashboard
    url: https://app.outerlayer.ai
  - description: Local dev server
    url: http://localhost:3002